OnTheMap
On the map is an app with a map that shows information posted by other students. The map contain pins that show the location where other students have reported studying. By tapping on the pin users can see a URL for something the student finds interesting. The user also able to add their own data by posting a string that can be geocoded to a location, and a URL.